:construction:  Added Plots class, which will be responsible for the plots
It's still a WIP, but the Plots class is abstract and needs to be extended by concrete classes actually implementing plotting
procedures. The only method required is get_methods(), which return a dict conatining an human-friendly name and the reference to the
method for each plotting procedure defined by the class. That dict will be used to populate dynamically the GUI. I implemented three
plotting procedures for CrossValidationPlots, all of them receiving only a CrossValidation object.