:boom:  Changing return types in RunCalculations class
I'm changing the return type of RunCalculations methods in order to be able to get back the CrossValidation (or ExternalValidation)
object, being then able to pass it to Plots. I'm trying to avoid breaking things, adding an optional parameter to toggle the new
behavior in each case, but it's obviouslly impossible to avoid everything, so I'm marking this as breaking changes.